+++ R/check-built.R
@@ -33,7 +33,7 @@ check_missing_images <- function(pkg, src_path, dst_path) {
     cli::cli_inform(c(
       x = "Missing alt-text in {.file {path_rel(src_path, pkg$src_path)}}",
       set_names(problems, "*"),
-      i = "Learn more in {.vignette accessibility}."
+      i = "Learn more in {.vignette pkgdown::accessibility}."
     ))
   }
 }