+++ tests/testthat/test-check.R
@@ -1,31 +1,39 @@
 test_that("sitrep complains about BS3", {
-  pkg <- local_pkgdown_site(test_path("assets/open-graph"), meta = "
-    template:
-      bootstrap: 3
-  ")
+  pkg <- local_pkgdown_site(
+    test_path("assets/open-graph"),
+    list(template = list(bootstrap = 3))
+  )
   expect_snapshot(pkgdown_sitrep(pkg))
 })
 
 test_that("sitrep reports all problems", {
-  pkg <- local_pkgdown_site(test_path("assets/reference"), meta = "
-    reference:
-     - title: Title
-       contents: [a, b, c, e]
-  ")
+  pkg <- local_pkgdown_site(
+    test_path("assets/reference"),
+    list(reference = list(
+      list(title = "Title", contents = c("a", "b", "c", "e"))
+    ))
+  )
+  
   expect_snapshot(pkgdown_sitrep(pkg))
 })
 
 test_that("checks fails on first problem", {
-  pkg <- local_pkgdown_site(test_path("assets/reference"), meta = "
-    reference:
-     - title: Title
-       contents: [a, b, c, e]
-  ")
+  pkg <- local_pkgdown_site(
+    test_path("assets/reference"),
+    list(reference = list(
+      list(title = "Title", contents = c("a", "b", "c", "e"))
+    ))
+  )
+  
   expect_snapshot(check_pkgdown(pkg), error = TRUE)
 })
 
 test_that("both inform if everything is ok", {
-  pkg <- test_path("assets/open-graph")
+  pkg <- local_pkgdown_site(
+    meta = list(url = "https://example.com"),
+    desc = list(URL = "https://example.com")
+  )
+
   expect_snapshot({
     pkgdown_sitrep(pkg)
     check_pkgdown(pkg)
@@ -42,4 +50,26 @@ test_that("check_urls reports problems", {
   # URL only in the pkgdown config
   pkg <- test_path("assets/cname")
   expect_snapshot(check_urls(pkg), error = TRUE)
+})
+
+# check favicons --------------------------------------------------------------
+
+test_that("check_favicons reports problems", {
+  pkg <- local_pkgdown_site()
+
+  # no logo no problems
+  expect_no_error(check_favicons(pkg))
+
+  # logo but no favicons
+  file_touch(path(pkg$src_path, "logo.svg"))
+  expect_snapshot(check_favicons(pkg), error = TRUE)
+  
+  # logo and old favicons
+  dir_create(path_favicons(pkg))
+  file_touch(path(path_favicons(pkg), "favicon.ico"), Sys.time() - 86400)
+  expect_snapshot(check_favicons(pkg), error = TRUE)
+
+  # logo and new favicons
+  file_touch(path(path_favicons(pkg), "favicon.ico"), Sys.time() + 86400)
+  expect_no_error(check_favicons(pkg))
 })