+++ tests/testthat/test-package.R
@@ -15,9 +15,11 @@ test_that("package_vignettes() doesn't trip over directories", {
 })
 
 test_that("check_bootstrap_version() allows 3, 4 (with warning), and 5", {
-  expect_equal(check_bootstrap_version(3), 3)
-  expect_warning(expect_equal(check_bootstrap_version(4), 5))
-  expect_equal(check_bootstrap_version(5), 5)
+  pkg <- local_pkgdown_site()
+
+  expect_equal(check_bootstrap_version(3, pkg), 3)
+  expect_snapshot(expect_equal(check_bootstrap_version(4, pkg), 5))
+  expect_equal(check_bootstrap_version(5, pkg), 5)
 })
 
 test_that("check_bootstrap_version() gives informative error otherwise", {
@@ -94,4 +96,27 @@ test_that("read_meta() errors gracefully if _pkgdown.yml failed to parse", {
     error = TRUE,
     transform = function(x) gsub(pkg$src_path, "<src>", x, fixed = TRUE)
   )
+})
+
+# lifecycle ---------------------------------------------------------------
+
+test_that("can extract lifecycle badges from description", {
+  rd_desc <- rd_text(
+    paste0("\\description{", lifecycle::badge("deprecated"), "}"),
+    fragment = FALSE
+  )
+  rd_param <- rd_text(
+    paste0("\\arguments{\\item{pkg}{", lifecycle::badge("deprecated"), "}}"),
+    fragment = FALSE
+  )
+
+  expect_equal(extract_lifecycle(rd_desc), "deprecated")
+  expect_equal(extract_lifecycle(rd_param), NULL)
+})
+
+test_that("malformed figures fail gracefully", {
+  rd_lifecycle <- function(x) extract_lifecycle(rd_text(x))
+
+  expect_null(rd_lifecycle("{\\figure{deprecated.svg}}"))
+  expect_null(rd_lifecycle("{\\figure{}}"))
 })