+++ tests/testthat/test-render.R
@@ -1,39 +1,5 @@
-test_that("check_bslib_theme() works", {
-  pkg <- as_pkgdown(test_path("assets/reference"))
-  expect_equal(check_bslib_theme("default", pkg, bs_version = 4), "default")
-  expect_equal(check_bslib_theme("lux", pkg, bs_version = 4), "lux")
-  expect_snapshot(error = TRUE, {
-    check_bslib_theme("paper", pkg, bs_version = 4)
-    check_bslib_theme("paper", pkg, bs_version = 4, field = c("template", "preset"))
-  })
-})
-
-test_that("get_bslib_theme() works with template.bslib.preset", {
-  pkg <- local_pkgdown_site(test_path("assets/site-empty"), '
-    template:
-      bootstrap: 5
-      bslib:
-        preset: shiny
-        enable-shadows: true
-  ')
-
-  expect_equal(get_bslib_theme(pkg), "shiny")
-  expect_no_error(bs_theme(pkg))
-
-  pkg <- local_pkgdown_site(test_path("assets/site-empty"), '
-    template:
-      bootstrap: 5
-      bslib:
-        preset: lux
-        enable-shadows: true
-  ')
-
-  expect_equal(get_bslib_theme(pkg), "lux")
-  expect_no_error(bs_theme(pkg))
-})
-
 test_that("capture data_template()", {
-  pkg <- as_pkgdown(test_path("assets/site-empty"))
+  pkg <- local_pkgdown_site()
   data <- data_template(pkg)
   data$year <- "<year>"
   data$footer$right <- gsub(packageVersion("pkgdown"), "{version}", data$footer$right, fixed = TRUE)
@@ -41,13 +7,15 @@ test_that("capture data_template()", {
 })
 
 test_that("can include text in header, before body, and after body", {
-  pkg <- local_pkgdown_site(test_path("assets/site-empty"), '
-    template:
-      includes:
-        in_header: <test>in header</test>
-        before_body: <test>before body</test>
-        after_body: <test>after body</test>
-  ')
+  pkg <- local_pkgdown_site(meta = list(
+    template = list(
+      includes = list(
+        in_header = "<test>in header</test>",
+        before_body = "<test>before body</test>",
+        after_body = "<test>after body</test>"
+      )
+    )
+  ))
 
   expect_named(
     data_template(pkg)$includes,
@@ -71,15 +39,17 @@ test_that("can include text in header, before body, and after body", {
 })
 
 test_that("check_opengraph validates inputs", {
-  check_open_graph_ <- function(...) {
-    check_open_graph(..., path = "_pkgdown.yml")
+  data_open_graph_ <- function(x) {
+    pkg <- local_pkgdown_site(meta = list(template = list(opengraph = x)))
+    data_open_graph(pkg)
+    invisible()
   }
 
   expect_snapshot(error = TRUE, {
-    check_open_graph_(list(foo = list()), )
-    check_open_graph_(list(foo = list(), bar = list()))
-    check_open_graph_(list(twitter = 1))
-    check_open_graph_(list(twitter = list()))
-    check_open_graph_(list(image = 1))
+    data_open_graph_(list(foo = list()))
+    data_open_graph_(list(foo = list(), bar = list()))
+    data_open_graph_(list(twitter = 1))
+    data_open_graph_(list(twitter = list()))
+    data_open_graph_(list(image = 1))
   })
 })