+++ tests/testthat/test-templates.R
@@ -92,7 +92,6 @@ test_that("BS5 templates have main + aside", {
   }
 })
 
-
 # Bootstrap version resolution --------------------------------------------
 test_that("Bootstrap version in template package under `template.bootstrap`", {
   path_template_package <- local_pkgdown_template_pkg(
@@ -149,64 +148,41 @@ test_that("Valid local Bootstrap version masks invalid template package", {
 
 # Bootstrap theme resolution ----------------------------------------------
 test_that("Finds Bootstrap theme in all the places", {
-  pkg_sketchy <- local_pkgdown_site(meta = '
-    template:
-      bslib:
-        preset: sketchy
-        version: 5
-  ')
-  pkg_superhero <- local_pkgdown_site(meta = '
-    template:
-      bslib:
-        preset: superhero
-        version: 5
-  ')
-  pkg_cosmo <- local_pkgdown_site(meta = '
-    template:
-      bootstrap: 5
-      bootswatch: cosmo
-  ')
-
-  pkg_yeti <- local_pkgdown_site(meta = '
-    template:
-      bootstrap: 5
-      params:
-        bootswatch: yeti
-  ')
-
+  pkg_sketchy <- local_pkgdown_site(meta = 
+    list(template = list(bslib = list(preset = "sketchy", version = 5)))
+  )
+  pkg_cosmo <- local_pkgdown_site(meta = 
+    list(template = list(bootstrap = 5, bootswatch = "cosmo"))
+  )
+  pkg_yeti <- local_pkgdown_site(meta = 
+    list(template = list(bootstrap = 5, params = list(bootswatch = "yeti")))
+  )
+  
   expect_equal(get_bslib_theme(pkg_sketchy), "sketchy")
-  expect_equal(get_bslib_theme(pkg_superhero), "superhero")
   expect_equal(get_bslib_theme(pkg_cosmo), "cosmo")
   expect_equal(get_bslib_theme(pkg_yeti), "yeti")
 })
 
 test_that("Warns when Bootstrap theme is specified in multiple locations", {
-  pkg <- local_pkgdown_site(meta = '
-    template:
-      bootstrap: 5
-      bootswatch: cerulean
-      bslib:
-        preset: flatly
-        bootswatch: lux
-      params:
-        bootswatch: darkly
-  ')
-
-  expect_snapshot(
-    get_bslib_theme(pkg)
-  )
+  pkg <- local_pkgdown_site(meta = list(
+    template = list(
+      bootstrap = 5,
+      bootswatch = "cerulean",
+      bslib = list(preset = "flatly", bootswatch = "lux"),
+      params = list(bootswatch = "darkly")
+    )
+  ))
+   expect_snapshot(get_bslib_theme(pkg))
 })
 
 test_that("Doesn't warn when the same Bootstrap theme is specified in multiple locations", {
-  pkg <- local_pkgdown_site(meta = '
-    template:
-      bootswatch: cerulean
-      bslib:
-        preset: cerulean
-  ')
-
-  expect_equal(
-    expect_silent(get_bslib_theme(pkg)),
-    "cerulean"
+  pkg <- local_pkgdown_site(meta = 
+    list(template = list(
+      bootstrap = 5,
+      bootswatch = "cerulean",
+      bslib = list(preset = "cerulean")
+    ))
   )
+
+  expect_equal(expect_silent(get_bslib_theme(pkg)), "cerulean")
 })